backend/auth/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
sqlalchemy